extract cli, add support for function jobs
