+++ deno.json
@@ -1,13 +1,25 @@
 {
-  "version": "0.4.0",
   "name": "@polyseam/cronx",
+
+  "version": "0.4.0",
+
+  "description": "Cross-Platform CLI and Library for working with Cron Jobs",
+
+  "license": "Apache-2.0",
+
+  "exports": {
+    "./nlp": "./src/nlp.ts",
+    "./cron": "./src/cron.ts",
+    "./cronx": "./src/cronx.ts"
+  },
+
   "tasks": {
     "run": "deno run --unstable-cron --allow-run main.ts",
+
     "checks": {
       "command": "deno lint && deno check . && deno fmt --check"
     },
-    "description": "Cross-Platform CLI and Library for working with Cron Jobs",
-    "license": "Apache-2.0",
+
     "archive-all": {
       "dependencies": [
         "archive-win-amd64",
@@ -109,8 +121,5 @@
     "@std/testing": "jsr:@std/testing@^1.0.9",
     "cconsole": "./src/cconsole.ts",
     "src/": "./src/"
-  },
-  "exports": {
-    "./nlp": "./src/nlp.ts"
   }
 }
